requirements.txt
numpy>=1.9.2
pandas
PyWavelets>=0.4.0
pykwalify>=1.6.0
six>=1.10.0
scikit-learn==1.0.2
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pedantic>=1.8.0,<2.0.0
tensorflow-cpu==2.9.2
gunicorn==20.0.4
python-multipart==0.0.5
SimpleITK>=0.9.1
pyradiomics==v3.0.1
